[MaL] renamed and moved FunctionHandles
