+++ MathLib/ODE/CVodeSolver.h
@@ -13,6 +13,7 @@
 #include "BaseLib/ConfigTree.h"
 
 #include "OdeSolverTypes.h"
+#include "FunctionHandles.h"
 
 namespace MathLib
 {
@@ -33,7 +34,7 @@ protected:
 	void setTolerance(double const* const abstol, const double reltol);
 	void setTolerance(const double abstol, const double reltol);
 
-	void setFunction(std::unique_ptr<FunctionHandles>&& f);
+	void setFunction(std::unique_ptr<detail::FunctionHandles>&& f);
 
 	void setIC(const double t0, double const* const y0);
