+++ MathLib/ODE/FunctionHandles.h
@@ -16,14 +16,31 @@ namespace MathLib
 {
 namespace detail
 {
+class FunctionHandles
+{
+public:
+	virtual bool call(const double t, double const* const y,
+	                  double* const ydot) = 0;
+	virtual bool callJacobian(const double t,
+	                          double const* const y,
+	                          double* const ydot,
+	                          double* const jac) = 0;
+
+	virtual bool hasJacobian() const = 0;
+
+	virtual unsigned getNumEquations() const = 0;
+
+	virtual ~FunctionHandles() = default;
+};
+
 /// Function handles for N equations.
 template <unsigned N>
-struct Handles : public MathLib::FunctionHandles
+struct FunctionHandlesImpl : FunctionHandles
 {
 	using Function = MathLib::Function<N>;
 	using JacobianFunction = MathLib::JacobianFunction<N>;
 
-	Handles(Function& f, JacobianFunction& df) : f(f), df(df) {}
+	FunctionHandlesImpl(Function& f, JacobianFunction& df) : f(f), df(df) {}
 	bool call(const double t, const double* const y,
 	          double* const ydot) override
 	{
